Update to be compatible with 117
The map position messages have changed, so bots will need to update
to their version of ropeybot to work with anyone running 117. This code
won't work with 116 or earlier, so don't update bots until 117 is out.

Also remove the CoordObject type that's now in bc-stubs as
ChatRoomMapPos: update code to use that instead of CoordObject.